+++ CHANGELOG.md
@@ -4,6 +4,18 @@
 
 | **Date**     | **Author**    | **Branch** |
 |--------------|---------------|------------|
+| [2025-10-02] | [Igee Theron] | [main]     |
+
+#### Added
+- Implemented login system with remember me logic
+- Implemented depot and pricing API
+- Implemented depot model, service and repository
+
+---
+## ChangeLog Entry
+
+| **Date**     | **Author**    | **Branch** |
+|--------------|---------------|------------|
 | [2025-10-01] | [Igee Theron] | [main]     |
 
 #### Added